requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
dataclasses==0.6
fastapi==0.61.0
h11==0.9.0
httptools==0.1.1
idna==2.6
Inject==4.3.1
keyring==10.6.0
keyrings.alt==3.0
pycrypto==2.6.1
pydantic==1.6.1
pygobject==3.26.1
PyMySQL==0.10.0
python-dotenv==0.14.0
pyxdg==0.25
requests==2.24.0
SecretStorage==2.3.1
six==1.11.0
soupsieve==2.0.1
starlette==0.13.6
tabelog-scraper==1.2
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
